@@ -1,6 +1,6 @@
 module github.com/google/googet/v2
 
-go 1.13
+go 1.16
 
 require (
 	cloud.google.com/go/storage v1.15.0
